finalized aqiff
