+++ app/src/main/res/layout/add_listing.xml
@@ -76,7 +76,9 @@
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.498"
-        app:layout_constraintStart_toStartOf="parent" />
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/imageView1"
+        app:layout_constraintVertical_bias="0.69" />
 
     <TextView
         android:id="@+id/itemName0"
@@ -161,14 +163,12 @@
 
     <ImageView
         android:id="@+id/imageView1"
-        android:layout_width="67dp"
-        android:layout_height="67dp"
+        android:layout_width="62dp"
+        android:layout_height="54dp"
         android:layout_marginTop="20dp"
-        app:layout_constraintBottom_toTopOf="@+id/submit"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/spinner1"
-        app:layout_constraintVertical_bias="0.017"
         app:srcCompat="@drawable/additem" />
 
     <Spinner
