+++ app/src/main/res/layout/interestscreen.xml
@@ -83,20 +83,20 @@
         app:layout_constraintTop_toBottomOf="@+id/btnSwipeJer" />
 
     <Button
-        android:id="@+id/button"
+        android:id="@+id/menfashion"
         android:layout_width="125dp"
         android:layout_height="wrap_content"
         android:layout_marginTop="100dp"
         android:background="@drawable/rounded_button"
         android:text="Men Fashion"
         android:textColor="#FFFFFF"
-        app:layout_constraintEnd_toStartOf="@+id/button2"
+        app:layout_constraintEnd_toStartOf="@+id/womenfashion"
         app:layout_constraintHorizontal_bias="0.5"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/editText" />
 
     <Button
-        android:id="@+id/button2"
+        android:id="@+id/womenfashion"
         android:layout_width="125dp"
         android:layout_height="wrap_content"
         android:layout_marginTop="100dp"
@@ -105,34 +105,34 @@
         android:textColor="#FFFFFF"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.5"
-        app:layout_constraintStart_toEndOf="@+id/button"
+        app:layout_constraintStart_toEndOf="@+id/menfashion"
         app:layout_constraintTop_toBottomOf="@+id/editText" />
 
     <Button
-        android:id="@+id/button3"
+        android:id="@+id/electronics"
         android:layout_width="125dp"
         android:layout_height="wrap_content"
         android:layout_marginTop="36dp"
         android:background="@drawable/rounded_button"
         android:text="Electronics"
         android:textColor="#FFFFFF"
-        app:layout_constraintEnd_toStartOf="@+id/button4"
+        app:layout_constraintEnd_toStartOf="@+id/gadgets"
         app:layout_constraintHorizontal_bias="0.5"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/button" />
+        app:layout_constraintTop_toBottomOf="@+id/menfashion" />
 
     <Button
-        android:id="@+id/button4"
+        android:id="@+id/gadgets"
         android:layout_width="125dp"
         android:layout_height="wrap_content"
         android:layout_marginTop="36dp"
         android:background="@drawable/rounded_button"
-        android:text="Gadget"
+        android:text="Gadgets"
         android:textColor="#FFFFFF"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.5"
-        app:layout_constraintStart_toEndOf="@+id/button3"
-        app:layout_constraintTop_toBottomOf="@+id/button2" />
+        app:layout_constraintStart_toEndOf="@+id/electronics"
+        app:layout_constraintTop_toBottomOf="@+id/womenfashion" />
 
     <ImageButton
         android:id="@+id/imageButton2"
@@ -150,8 +150,21 @@
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.555"
         app:layout_constraintStart_toStartOf="parent"
-        app:srcCompat="@drawable/ic_addbut"
-        android:onClick="addItemPage"/>
+        app:srcCompat="@drawable/ic_addbut"/>
+
+    <ImageButton
+        android:id="@+id/searchItem"
+        android:layout_width="37dp"
+        android:layout_height="37dp"
+        android:layout_marginTop="156dp"
+        android:layout_weight="1"
+        android:adjustViewBounds="true"
+        android:background="?attr/selectableItemBackgroundBorderless"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.085"
+        app:layout_constraintStart_toEndOf="@+id/editText"
+        app:layout_constraintTop_toTopOf="parent"
+        app:srcCompat="@drawable/ic_action_name" />
 
 
 </androidx.constraintlayout.widget.ConstraintLayout>
