+++ app/src/main/res/values/strings.xml
@@ -14,7 +14,7 @@
         <item>Men Fashion</item>
         <item>Women Fashion</item>
         <item>Electronics</item>
-        <item>Mobile Gadgets</item>
+        <item>Gadgets</item>
     </string-array>
 
 </resources>
